refactor(ProdcutsGrid): 공통 컴포넌트로 분리
- ThemeProductPage, TredingGifts 를 type props 로 구분
